optimized spline point placement
This script is meant to find the optimal placement of an extra spline point for the LV3 nose cone. The curvature of the LD-Haack shape near the tip creates a large error when using evenly distributed sample points to approximate the shape using spline curves. This script compares different placements of an extra sample point, to minimize that error.