made a model to compensate for fin mistake; rearranged ORK files
It turns out that the fins - which no-one was given a chance to review
before they were sent out for manufacturing - have a major flaw: they
are about 70% as long as they're supposed to be. (wow!) This is in
addition to the concerns about the root fillet, which were also closed
to review before machining.

This means that the CoP is now much farther forward than designed. To
compensate for this, we need to add about 6 kg of lead to the end of the
nose cone! (This moves the CoM further forward.) Otherwise, the rocket will spin around and crash.