Modified the fin sweep to match OpenRocket
Normally, sweep angle is defined as the angle (from perpendicular) of the line following the middle of the airfoil.
(That's the "correct" definition, and is more-or-less what was used in the model.)
In OpenRocket, it's defined as the angle (from perpendicular) of the leading edge.
(Alternatively, OR lets you define the sweep length of the fin: how far the start of the tip is behind the start of the root.)

I changed it to be defined by the sweep length and fixed a problem that made with extruding the rear truss.