cleaned up the document and fixed LaTeX stuff
The LaTeX equations are now rendered as part of an R plot.
It's a little bit of a ghetto solution, but my system really doesn't
want to render plots and latex in the same document.
I might write a short gist or something about automated documents with
R, since I've spent a fair bit of time on it.
(in addition to what I've done with this document)