mock-up of poster layout; added poster resources
If someone could look at the layout (doc/poster/posterLayout_45x40in.jpg)
and tell me what you think, that would be great.
I'm going to start working on it, and hopefully noone has very drastic
changes to make to it.

I added some QR code EPS files and a bash script to regenerate them.
They are in there so people easily can email us and check the PSAS site and the
repo.