completed fin can layup assembly
Major things to note:
* I changed the "T" block for the sled to give some room for the
adhesive sheet and some spring washers.
* I added a baseplate (intended to be made out of MDF). It interfaces
with the female ring to align the module to the fins in the
theta-direction. It also has 3 holes for leveling feet and a center hole
that we could potentially run a 1/4" rod through (with the intent of
attaching that rod to a spider on the other end, so we can fix the
module to the baseplate).
* I put in some PTFE tubes that have been cut in half. These sit between
the leading/trailing edges and the jigs. This way, we can still controll
the axial position of the fins while not risking any damage to the
leading/trailing edges.