Found the problem in the DATCOM model! (kinda)
It's something to do with the station-wise definition of the body not
working at high Mach numbers. DATCOM gives normal outputs for the model
at low Mach numbers, but it poops out at high (~1.5) Mach numbers.

It also doesn't seem to like the altitudes being specified in inches,
which is annoying. I'm not sure if changing the dimensions into feet
would help. It would be a pain to do though... I'm not sure if it would
be worth it to write a script that just divides all the numbers by
12, lol.

I'm also not sure if it would help at all to reformulate the body
definition into an ogive and a cylinder. The DATCOM manual is a little
unclear on what exactly counts as an ogive. I would guess that at a 1:5
aspect ratio, there's not much practical difference between a von Karman
ogive and whatever rando ogive. I would also guess that there's more
data on known nosecone geometries than on arbitrary station-wise bodies.