corrected rounding/legacy errors in DATCOM input
I used the `y()` function in
`sim/OpenFOAM/LV3_LD-Haack/noseSplineOptimizer.R` (modified to units of
inches, of course) to correct rounding errors in the postions and areas
of the body stations.

I compared to the OpenRocket model to double-check the
dimensions/locations of the fins and the overall length of the rocket.

I changed some parameters of the airfoil to correspond to the new 1/8"
fins, rather than the old 1/4" fins.

TODO:
* correct the airfoil dimensions (see Vol.1 pg.154 of the Digital
DATCOM manual)
    * correct X_t to correspond to the angle _actually_ created by the
    chamfer endmill, rather than the designed angle.
    * same for X_f
* Mess with the FLTCON parameters to match the altitudes/speeds/AoAs
predicted by OpenRocket.