created a clamp that holds the brackets down
This hooks onto the fin jig plates. A nut is trapped in the obvious
spot. A bolt threads through to apply clamping force to the brackets in
the fin can layup. (I'll try to remember to add this into the assembly
at some point.)

This is meant to be a 3D printed part. The layers should be coplanar
with the sides, for maximum strength. It should have a thick wall (~2 mm)
and maybe 50% infill. (We want this part to bear around 20 pounds.)