added fudge mass to match simulation to cert card
When we physically measured the CoM of the rocket, we found that it was
significantly further forward from the predicted CoM. So, some of my
bookkeeping  was off. (This shouldn't be too surprising, since I don't
think we ever got an updated measurement for the recovery system.)
But... the fact that the fudge mass is outside of the nosecone indicates
that something towards the tail overreported its weight. I'd assume it's
the motor.